remove DatasetOutputPlugin and replace by a direct implementation of IntoOutputSystem for ToDataset
