+++ src/physics/mod.rs
@@ -19,7 +19,6 @@ use crate::communication::Communicator;
 use crate::domain::ExchangeDataPlugin;
 use crate::io::input::DatasetInputPlugin;
 use crate::io::output::AttributeOutputPlugin;
-use crate::io::output::DatasetOutputPlugin;
 use crate::io::output::OutputPlugin;
 use crate::mass::Mass;
 use crate::named::Named;
@@ -55,13 +54,13 @@ impl RaxiomPlugin for PhysicsPlugin {
         sim.add_plugin(ExchangeDataPlugin::<Position>::default())
             .add_plugin(ExchangeDataPlugin::<Velocity>::default())
             .add_plugin(ExchangeDataPlugin::<Mass>::default())
-            .add_plugin(OutputPlugin(DatasetOutputPlugin::<Position>::default()))
-            .add_plugin(OutputPlugin(DatasetOutputPlugin::<Velocity>::default()))
-            .add_plugin(OutputPlugin(DatasetOutputPlugin::<Mass>::default()))
+            .add_plugin(OutputPlugin::<Position>::default())
+            .add_plugin(OutputPlugin::<Velocity>::default())
+            .add_plugin(OutputPlugin::<Mass>::default())
             .add_plugin(DatasetInputPlugin::<Position>::default())
             .add_plugin(DatasetInputPlugin::<Velocity>::default())
             .add_plugin(DatasetInputPlugin::<Mass>::default())
-            .add_plugin(OutputPlugin(AttributeOutputPlugin::<Time>::default()))
+            .add_plugin(OutputPlugin::<AttributeOutputPlugin<Time>>::default())
             .add_plugin(CommunicationPlugin::<ShouldExit>::new(
                 CommunicationType::AllGather,
             ))
